I translated into Japanese.
I translated it into Japanese from scratch based on "TsunKitTranslations/en/kcnav.json".

The translation was done based on my own experience of using it and the following pages.
https://wikiwiki.jp/kancolle/%E3%83%AB%E3%83%BC%E3%83%88%E5%88%86%E5%B2%90#how_to_kcnav
(By the way, I wrote this page・。・ｖ)

The pioneer has translated the ship class list in "Ship Classes" into Japanese, but I think it is fine to leave it in English.
One of the reasons is that the search is done using the English ship class name.
The second reason is that Japanese admirals often mistake "潜空"(SSV:潜水空母) and "潜母"(AS:潜水母艦). Abbreviations of ship types in Japanese are ambiguous and should be avoided.
Instead, I suggest that each ship type in the "Ship Classes" have a tooltip with a Japanese translation. I've already added a tip for the Japanese translation of the ship classes in the json I pull request.

Also, I couldn't come up with a good Japanese translation, so I added new explanatory tooltips to "mapMenu", "panelStatsFSRadar", etc.

I apologize for adding tooltips in places where there were originally no tips.

I writed this sentence using a translating machine. Apologize if the rude expression.